requirements.txt
# what all libraries to to installed
streamlit
google-generativeai
python.dotenv